feat: individual file for History Tests
