+++ QuantConnect.IEX.Tests/IEXDataQueueHandlerTests.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -216,157 +216,5 @@ namespace QuantConnect.IEX.Tests
                 true,
                 false);
         }
-
-        #region History provider tests
-
-        /// <summary>
-        /// Provides test parameters for the TestMethod.
-        /// </summary>
-        /// <remarks>
-        /// The test parameters include valid and invalid combinations of input data.
-        /// </remarks>
-        internal static IEnumerable<TestCaseData> TestParameters
-        {
-            get
-            {
-                // Valid parameters
-                yield return new TestCaseData(Symbols.SPY, Resolution.Daily, TickType.Trade, TimeSpan.FromDays(15), true)
-                    .SetDescription("Valid parameters - Daily resolution, 15 days period.")
-                    .SetCategory("Valid");
-
-                yield return new TestCaseData(Symbols.SPY, Resolution.Minute, TickType.Trade, TimeSpan.FromDays(5), true)
-                    .SetDescription("Valid parameters - Minute resolution, 5 days period.")
-                    .SetCategory("Valid");
-
-                // Invalid resolution - empty result
-                yield return new TestCaseData(Symbols.SPY, Resolution.Tick, TickType.Trade, TimeSpan.FromSeconds(15), false)
-                    .SetDescription("Invalid resolution - Tick resolution, 15 seconds period.")
-                    .SetCategory("Invalid");
-
-                yield return new TestCaseData(Symbols.SPY, Resolution.Second, TickType.Trade, Time.OneMinute, false)
-                    .SetDescription("Invalid resolution - Second resolution, 1 minute period.")
-                    .SetCategory("Invalid");
-
-                yield return new TestCaseData(Symbols.SPY, Resolution.Hour, TickType.Trade, Time.OneDay, false)
-                    .SetDescription("Invalid resolution - Hour resolution, 1 day period.")
-                    .SetCategory("Invalid");
-
-                // Invalid period - empty result
-                yield return new TestCaseData(Symbols.SPY, Resolution.Minute, TickType.Trade, TimeSpan.FromDays(45), false)
-                    .SetDescription("Invalid period - Beyond 30 days, Minute resolution.")
-                    .SetCategory("Invalid");
-
-                yield return new TestCaseData(Symbols.SPY, Resolution.Daily, TickType.Trade, TimeSpan.FromDays(-15), false)
-                    .SetDescription("Invalid period - Date in the future, Daily resolution.")
-                    .SetCategory("Invalid");
-
-                // Invalid data type - empty result
-                yield return new TestCaseData(Symbols.SPY, Resolution.Daily, TickType.Quote, TimeSpan.FromDays(15), false)
-                    .SetDescription("Invalid data type - Daily resolution, QuoteBar data type.")
-                    .SetCategory("Invalid");
-
-                // Invalid security type, no exception, empty result
-                yield return new TestCaseData(Symbols.EURUSD, Resolution.Daily, TickType.Trade, TimeSpan.FromDays(15), false)
-                    .SetDescription("Invalid security type - EURUSD symbol, Daily resolution.")
-                    .SetCategory("Invalid");
-            }
-        }
-
-        [Test, TestCaseSource(nameof(TestParameters))]
-        public void IEXCouldGetHistory(Symbol symbol, Resolution resolution, TickType tickType, TimeSpan period, bool received)
-        {
-            var slices = GetHistory(symbol, resolution, tickType, period);
-
-            if (!received)
-            {
-                Assert.IsEmpty(slices);
-                return;
-            }
-
-            Assert.IsNotEmpty(slices);
-
-            foreach (var slice in slices)
-            {
-                foreach (var data in slice)
-                {
-                    AssertTradeBar(symbol, resolution, data.Value, data.Key);
-                }
-            }
-
-            // And are ordered by time
-            Assert.That(slices, Is.Ordered.By("Time"));
-        }
-
-        internal static void AssertTradeBar(Symbol expectedSymbol, Resolution resolution, BaseData baseData, Symbol actualSymbol = null)
-        {
-            if (actualSymbol != null)
-            {
-                Assert.That(actualSymbol, Is.EqualTo(expectedSymbol));
-            }
-
-            Assert.That(baseData.DataType, Is.EqualTo(MarketDataType.TradeBar));
-
-            var tradeBar = baseData as TradeBar;
-            Assert.IsNotNull(tradeBar);
-            Assert.Greater(tradeBar.Open, 0);
-            Assert.Greater(tradeBar.High, 0);
-            Assert.Greater(tradeBar.Close, 0);
-            Assert.Greater(tradeBar.Low, 0);
-            Assert.That(tradeBar.Period.ToHigherResolutionEquivalent(true), Is.EqualTo(resolution));
-        }
-
-        /// <summary>
-        /// Provides test data for scenarios involving an invalid symbol.
-        /// </summary>
-        /// <remarks>
-        /// The test case includes an attempt to create a symbol ("XYZ") with an invalid combination of SecurityType and Market.
-        /// </remarks>
-        public static IEnumerable<TestCaseData> InvalidSymbolTestCaseData
-        {
-            get
-            {
-                yield return new TestCaseData(Symbol.Create("XYZ", SecurityType.Equity, Market.FXCM), Resolution.Daily, TickType.Trade, TimeSpan.FromDays(15))
-                    .SetDescription("Invalid symbol - Attempt to create a symbol with an invalid combination of SecurityType and Market.")
-                    .SetCategory("Invalid");
-            }
-        }
-
-        [Test, TestCaseSource(nameof(InvalidSymbolTestCaseData))]
-        public void GetHistoryInvalidSymbolThrowException(Symbol symbol, Resolution resolution, TickType tickType, TimeSpan period)
-        {
-            Assert.Throws<Exception>(() => GetHistory(symbol, resolution, tickType, period));
-        }
-
-        private Slice[] GetHistory(Symbol symbol, Resolution resolution, TickType tickType, TimeSpan period)
-        {
-            var requests = new[] { CreateHistoryRequest(symbol, resolution, tickType, period) };
-
-            var slices = iexDataQueueHandler.GetHistory(requests, TimeZones.Utc).ToArray();
-            Log.Trace("Data points retrieved: " + iexDataQueueHandler.DataPointCount);
-            Log.Trace("tick Type: " + tickType);
-            return slices;
-        }
-
-        internal static HistoryRequest CreateHistoryRequest(Symbol symbol, Resolution resolution, TickType tickType, TimeSpan period)
-        {
-            var now = DateTime.UtcNow;
-
-            var dataType = LeanData.GetDataType(resolution, tickType);
-
-            return new HistoryRequest(now.Add(-period),
-                                       now,
-                                       dataType,
-                                       symbol,
-                                       resolution,
-                                       SecurityExchangeHours.AlwaysOpen(TimeZones.NewYork),
-                                       TimeZones.NewYork,
-                                       resolution,
-                                       true,
-                                       false,
-                                       DataNormalizationMode.Raw,
-                                       tickType);
-        }
-
-        #endregion
     }
 }
